Refactor: - Move Navigation and Screen to core package - Move SpecialDiceName to domain package - Move SpecialDice to domain package - Update OwnedSpecialDice to validate isSelected and count - Update SpecialDiceCard to use displayName of SpecialDiceName
