+++ app/src/main/java/pl/kazoroo/tavernFarkle/shop/domain/model/SpecialDice.kt
@@ -1,11 +1,15 @@
 package pl.kazoroo.tavernFarkle.shop.domain.model
 
 import androidx.annotation.DrawableRes
-import pl.kazoroo.tavernFarkle.shop.data.model.SpecialDiceName
 
 data class SpecialDice(
     val name: SpecialDiceName,
     val price: Int,
     @DrawableRes val image: Int,
     val chancesOfDrawingValue: List<Float>
-)
+) {
+    init {
+        require(chancesOfDrawingValue.size == 6) { "chancesOfDrawingValue must contain 6 elements" }
+        require(price > 0) { "price must be greater than 0" }
+    }
+}
