+++ app/src/main/java/pl/kazoroo/tavernFarkle/shop/domain/model/SpecialDiceName.kt
@@ -1,0 +1,6 @@
+package pl.kazoroo.tavernFarkle.shop.domain.model
+
+enum class SpecialDiceName(val displayName: String) {
+    ODD_DICE("Odd Dice"),
+    ALFONSES_DICE("Alfonse's Dice")
+}
